Add a third parameter to generate Absolute Route
The function Routing.generate can take a third boolen argument like the router service to generate absoluteUrl.
The function will also check if the requested url has a requirement on the "_scheme"
If the scheme is not the same as the current scheme, then the function will return absolute path to fit route params.