Simplified the Travis configuration
Builds for the different PHP versions are using the normal Composer
resolution.
Testing against a few extra Symfony versions is done only for a single
PHP version (5.6) and requires the full framework to avoid issues with
dependency stability when testing the dev version of Symfony.
This also enables testing on PHP 5.6 and HHVM.